Refactor: move sijoittelun-tulos cells to their own files
