+++ src/app/haku/[oid]/hakukohde/[hakukohde]/sijoittelun-tulokset/components/sijoittelun-tulos-styled-cell.tsx
@@ -1,0 +1,8 @@
+import { Box, styled } from '@mui/material';
+
+export const SijoittelunTulosStyledCell = styled(Box)(({ theme }) => ({
+  display: 'flex',
+  flexDirection: 'column',
+  rowGap: theme.spacing(1),
+  minWidth: '240px',
+}));
