+++ src/app/haku/[oid]/hakukohde/[hakukohde]/sijoittelun-tulokset/components/sijoittelun-tulos-table.tsx
@@ -6,30 +6,16 @@ import {
   makeCountColumn,
 } from '@/app/components/table/table-columns';
 import { ListTable } from '@/app/components/table/list-table';
-import {
-  IlmoittautumisTila,
-  SijoittelunHakemusEnriched,
-  SijoittelunTila,
-  VastaanottoTila,
-} from '@/app/lib/types/sijoittelu-types';
-import { useMemo, useState } from 'react';
-import {
-  Box,
-  Checkbox,
-  FormControlLabel,
-  styled,
-  Typography,
-} from '@mui/material';
-import { OphFormControl } from '@/app/components/form/oph-form-control';
-import { LocalizedSelect } from '@/app/components/localized-select';
-import { MaksunTila } from '@/app/lib/types/ataru-types';
-import { useHyvaksynnanEhdot } from '../hooks/useHyvaksynnanEhdot';
-import { OphInput } from '@/app/components/form/oph-input';
-import { toFormattedDateTimeString } from '@/app/lib/localization/translation-utils';
+import { SijoittelunHakemusEnriched } from '@/app/lib/types/sijoittelu-types';
+import { useMemo } from 'react';
 import {
   KeysMatching,
   ListTableColumn,
 } from '@/app/components/table/table-types';
+import { MaksuCell } from './maksu-cell';
+import { IlmoittautumisCell } from './ilmoittautumis-cell';
+import { VastaanOttoCell } from './vastaanotto-cell';
+import { SijoittelunTilaCell } from './sijoittelun-tila-cell';
 
 export const makeEmptyCountColumn = <T extends Record<string, unknown>>({
   title,
@@ -47,174 +33,6 @@ export const makeEmptyCountColumn = <T extends Record<string, unknown>>({
 });
 
 const TRANSLATIONS_PREFIX = 'sijoittelun-tulokset.taulukko';
-
-const StyledCell = styled(Box)(({ theme }) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  rowGap: theme.spacing(1),
-  minWidth: '240px',
-}));
-
-const showHyvaksyVarasijalta = (hakemus: SijoittelunHakemusEnriched) =>
-  hakemus.tila === SijoittelunTila.VARALLA ||
-  (hakemus.hyvaksyttyVarasijalta &&
-    [
-      SijoittelunTila.HYVAKSYTTY,
-      SijoittelunTila.PERUUNTUNUT,
-      SijoittelunTila.VARALLA,
-      SijoittelunTila.VARASIJALTA_HYVAKSYTTY,
-    ].includes(hakemus.tila));
-
-const SijoittelunTilaCell = ({
-  hakemus,
-}: {
-  hakemus: SijoittelunHakemusEnriched;
-}) => {
-  const { t, translateEntity } = useTranslations();
-  const { data: hyvaksynnanEhdot } = useHyvaksynnanEhdot();
-
-  const ehtoOptions = hyvaksynnanEhdot.map((ehto) => {
-    return { value: ehto.koodiArvo, label: translateEntity(ehto.nimi) };
-  });
-
-  const [ehdollinen, setEhdollinen] = useState<boolean>(
-    hakemus.ehdollisestiHyvaksyttavissa,
-  );
-  const [ehdollinenSyy, setEhdollinenSyy] = useState(
-    hakemus.ehdollisenHyvaksymisenEhtoKoodi,
-  );
-
-  return (
-    <StyledCell>
-      <span>{t(`sijoitteluntila.${hakemus.tila}`)}</span>
-      {showHyvaksyVarasijalta(hakemus) && (
-        <FormControlLabel
-          label={t('sijoittelun-tulokset.varasijalta')}
-          control={
-            <Checkbox
-              checked={hakemus.hyvaksyttyVarasijalta}
-              onChange={() => ''}
-            />
-          }
-        />
-      )}
-      <FormControlLabel
-        label={t('sijoittelun-tulokset.ehdollinen')}
-        control={
-          <Checkbox
-            checked={ehdollinen}
-            onChange={() => setEhdollinen(!ehdollinen)}
-          />
-        }
-      />
-      {ehdollinen && (
-        <LocalizedSelect
-          sx={{ maxWidth: '300px' }}
-          value={ehdollinenSyy}
-          onChange={(event) => setEhdollinenSyy(event.target.value)}
-          options={ehtoOptions}
-        />
-      )}
-      {ehdollinen && ehdollinenSyy === 'muu' && (
-        <Box>
-          <OphInput
-            value={hakemus.ehdollisenHyvaksymisenEhtoFI ?? ''}
-            onChange={() => ''}
-          />
-          <OphInput
-            value={hakemus.ehdollisenHyvaksymisenEhtoSV ?? ''}
-            onChange={() => ''}
-          />
-          <OphInput
-            value={hakemus.ehdollisenHyvaksymisenEhtoEN ?? ''}
-            onChange={() => ''}
-          />
-        </Box>
-      )}
-    </StyledCell>
-  );
-};
-
-const VastaanOttoCell = ({
-  hakemus,
-}: {
-  hakemus: SijoittelunHakemusEnriched;
-}) => {
-  const { t } = useTranslations();
-
-  const vastaanottotilaOptions = Object.values(VastaanottoTila).map((tila) => {
-    return { value: tila as string, label: t(`vastaanottotila.${tila}`) };
-  });
-
-  return (
-    <StyledCell>
-      <FormControlLabel
-        label={t('sijoittelun-tulokset.julkaistavissa')}
-        control={
-          <Checkbox checked={hakemus.julkaistavissa} onChange={() => ''} />
-        }
-      />
-      {hakemus.vastaanottoDeadline && (
-        <Typography>
-          {t('sijoittelun-tulokset.vastaanottoaikaraja')}:{' '}
-          {toFormattedDateTimeString(hakemus.vastaanottoDeadline)}
-        </Typography>
-      )}
-      <OphFormControl
-        label={t('sijoittelun-tulokset.hakijalle-naytetaan')}
-        sx={{ fontWeight: 400 }}
-        renderInput={({ labelId }) => (
-          <LocalizedSelect
-            labelId={labelId}
-            value={hakemus.vastaanottotila}
-            onChange={() => ''}
-            options={vastaanottotilaOptions}
-          />
-        )}
-      />
-    </StyledCell>
-  );
-};
-
-const IlmoittautumisCell = ({
-  hakemus,
-}: {
-  hakemus: SijoittelunHakemusEnriched;
-}) => {
-  const { t } = useTranslations();
-
-  const ilmoittautumistilaOptions = Object.values(IlmoittautumisTila).map(
-    (tila) => {
-      return { value: tila as string, label: t(`ilmoittautumistila.${tila}`) };
-    },
-  );
-
-  return (
-    <LocalizedSelect
-      value={hakemus.ilmoittautumisTila}
-      onChange={() => ''}
-      options={ilmoittautumistilaOptions}
-    />
-  );
-};
-
-const MaksuCell = ({ hakemus }: { hakemus: SijoittelunHakemusEnriched }) => {
-  const { t } = useTranslations();
-
-  const maksuntilaOptions = Object.values(MaksunTila).map((tila) => {
-    return { value: tila as string, label: t(`maksuntila.${tila}`) };
-  });
-
-  return hakemus.maksuntila ? (
-    <LocalizedSelect
-      value={hakemus.maksuntila}
-      onChange={() => ''}
-      options={maksuntilaOptions}
-    />
-  ) : (
-    <></>
-  );
-};
 
 export const SijoittelunTulosTable = ({
   hakemukset,
